override target
The "-target native" argument causes Zig to call
detectAbiAndDynamicLinker() and determine the dynamic-linker from the
ELF data of `/usr/bin/env`. Problem is, `env` is a 32-bit executable,
and the zig binary is 64-bit.

One other person hit this same issue but in this case their
`/usr/bin/env` was a link to a statically built busybox, meaning there
was no interpreter to find.

https://github.com/ziglang/zig/issues/14577

The real fix is probably to just have detectAbiAndDynamicLinker() use
hard-coded defaults on the `solaris` tag. But for now I'm going to try
to make progress by setting the target explicitly.